+++ src/allocator_api.rs
@@ -1,23 +1,23 @@
-#[cfg(feature = "nightly")]
+#[cfg(feature = "allocator_api")]
 pub use core::alloc::{AllocError, Allocator};
 
-#[cfg(not(feature = "nightly"))]
+#[cfg(not(feature = "allocator_api"))]
 use core::{
     alloc::Layout,
     fmt::{self, Display},
     ptr::NonNull,
 };
 
-#[cfg(all(feature = "nightly", feature = "alloc"))]
+#[cfg(all(feature = "allocator_api", feature = "alloc"))]
 pub use alloc::alloc::Global;
 
 /// Same as [`core::alloc::AllocError`], but stable.
 /// When nightly feature is enabled, this is re-export of [`core::alloc::AllocError`] instead.
-#[cfg(not(feature = "nightly"))]
+#[cfg(not(feature = "allocator_api"))]
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub struct AllocError;
 
-#[cfg(not(feature = "nightly"))]
+#[cfg(not(feature = "allocator_api"))]
 impl Display for AllocError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         f.write_str("memory allocation failed")
@@ -26,7 +26,7 @@ impl Display for AllocError {
 
 /// Same as [`core::alloc::Allocator`], but stable.
 /// When nightly feature is enabled, this is re-export of [`core::alloc::Allocator`] instead.
-#[cfg(not(feature = "nightly"))]
+#[cfg(not(feature = "allocator_api"))]
 pub unsafe trait Allocator {
     /// See [`core::alloc::Allocator::allocate`].
     fn allocate(&self, layout: Layout) -> Result<NonNull<[u8]>, AllocError>;
@@ -35,12 +35,27 @@ pub unsafe trait Allocator {
     unsafe fn deallocate(&self, ptr: NonNull<u8>, layout: Layout);
 }
 
+#[cfg(not(feature = "allocator_api"))]
+unsafe impl<A> Allocator for &'_ A
+where
+    A: Allocator,
+{
+    fn allocate(&self, layout: Layout) -> Result<NonNull<[u8]>, AllocError> {
+        A::allocate(*self, layout)
+    }
+
+    unsafe fn deallocate(&self, ptr: NonNull<u8>, layout: Layout) {
+        A::deallocate(*self, ptr, layout)
+    }
+}
+
 /// Same as [`alloc::alloc::Global`], but stable.
 /// When nightly feature is enabled, this is re-export of [`alloc::alloc::Global`] instead.
-#[cfg(all(not(feature = "nightly"), feature = "alloc"))]
+#[cfg(all(not(feature = "allocator_api"), feature = "alloc"))]
+#[derive(Clone, Copy, Debug, Default)]
 pub struct Global;
 
-#[cfg(all(not(feature = "nightly"), feature = "alloc"))]
+#[cfg(all(not(feature = "allocator_api"), feature = "alloc"))]
 unsafe impl Allocator for Global {
     fn allocate(&self, layout: Layout) -> Result<NonNull<[u8]>, AllocError> {
         if layout.size() == 0 {