+++ src/lib.rs
@@ -1,23 +1,134 @@
-//! Scoped-Arena provides arena-like allocator with optional scopes.
-//! Arena allocators are simple and provides ludicrously fast allocation.
 //!
-//! `Scope` allows storing values. When dropped `Scope` will `Drop` all stored values and free memory allocated by the scope.\
+//! Scoped-Arena provides arena allocator with explicit scopes.
+//!
+//! ## Arena allocation
+//!
+//! Arena allocators are simple and provides ludicrously fast allocation.\
+//! Basically allocation requires only increment of internal pointer in the memory block to alignment of allocated object and then to size of allocated object and that's it.\
+//! When memory block is exhausted arena will allocate new bigger memory block.\
+//! Then arena can be reset after all allocated objects are not used anymore, keeping only last memory block and reuse it.\
+//! After several warmup iterations the only memory block is large enough to handle all allocations until next reset.
+//!
+//!
+//! ### Example
+//!
+//! ```rust
+//! use scoped_arena::Scope;
+//!
+//! struct Cat {
+//!     name: String,
+//!     hungry: bool,
+//! }
+//!
+//! /// Create new arena with `Global` allocator.
+//! let mut scope = Scope::new();
+//!
+//! /// Construct a cat and move it to the scope.
+//! let cat: &mut Cat = scope.to_scope(Cat {
+//!     name: "Fluffy".to_owned(),
+//!     hungry: true,
+//! });
+//!
+//! // Now `cat` is a mutable reference bound to scope borrow lifetime.
+//!
+//! assert_eq!(&cat.name, "Fluffy");
+//! assert!(cat.hungry);
+//!
+//! cat.hungry = false;
+//!
+//! // This cat instance on scope will be automatically dropped when `scope` is dropped or reset.
+//! // It is impossible to reset before last usage of `cat`.
+//!
+//! // Next line will drop cat value and free memory occupied by it.
+//! scope.reset();
+//!
+//! // If there were more cats or any other objects put on scope they all would be dropped and memory freed.
+//! ```
+//!
+//! ## Scopes
+//!
+//! To reuse memory earlier this crates provides `Scope` with methods to create sub-`Scope`s.\
+//! When sub-`Scope` is reset or dropped it will `Drop` all stored values and free memory allocated by the scope and flush last of new allocated memory block into parent.\
+//! While objects allocated with parent `Scope` are unchanged and still valid.
+//!
 //! Well placed scopes can significantly reduce memory consumption.\
-//! Creating new scope is cheap and allocating within scope is as fast as allocating parent arena allocator.\
-//! Drop glue is stored in scope only for values that `needs_drop`. i.e. for `Copy` types storing value on scope is simply allocation + copy.
+//! For example if few function calls use a lot of dynamic memory but don't need it to be available in caller\
+//! they can be provided with sub-scope.\
+//! At the same time any memory allocated in parent scope stays allocated.
 //!
-//! Rust borrowing rules require additional layer of `ScopeProxy` to allow references to values on scope and sub-scopes to coexist.
+//! Creating sub-scope is cheap and allocating within sub-scope is as fast as allocating in parent scope.\
+//!
+//! ### Example
+//!
+//! ```rust
+//! use scoped_arena::{Scope, ScopeProxy};
+//!
+//!
+//! fn heavy_on_memory(mut scope: Scope<'_>, foobar: &String) {
+//!     for _ in 0 .. 42 {
+//!         let foobar: &mut String = scope.to_scope(foobar.clone());
+//!     }
+//!
+//!     // new `scope` is dropped here and drops all allocated strings and frees memory.
+//! }
+//!
+//! let mut scope = Scope::new();
+//!
+//! // Proxy is required to be friends with borrow checker.
+//! // Creating sub-scope must lock parent `Scope` from being used, which requires mutable borrow, but any allocation borrows `Scope`.
+//! // `Proxy` relaxes this a bit. `Proxy` borrows `Scope` mutably and tie allocated objects lifetime to scopes' borrow lifetime.
+//! // So sub-scope can borrow proxy mutably while there are objects allocated from it.
+//! let mut proxy = scope.proxy();
+//!
+//! let foobar: &mut String = proxy.to_scope("foobar".to_owned());
+//!
+//! // Make sub-scope for the call.
+//! heavy_on_memory(proxy.scope(), &*foobar);
+//!
+//! // If `heavy_on_memory` didn't trigger new memory object allocation in the scope,
+//! // sub-scope drop would rewind scope's internals to exactly the same state.
+//! // Otherwise last of new blocks will become current block in parent scope.
+//! //
+//! // Note that `foobar` is still alive.
+//!
+//! heavy_on_memory(proxy.scope(), &*foobar);
+//! heavy_on_memory(proxy.scope(), &*foobar);
+//! heavy_on_memory(proxy.scope(), &*foobar);
+//! heavy_on_memory(proxy.scope(), &*foobar);
+//!
+//! // Once peak memory consumption is reached, any number of `heavy_on_memory` calls would not require new memory blocks to be allocated.
+//! // Even `loop { heavy_on_memory(proxy.scope(), &*foobar) }` will settle on some big enough block.
+//! ```
+//!
+//! ## Dropping
+//!
+//! `to_scope` and `try_to_scope` methods store drop-glue for values that `needs_drop`.
+//! On reset or drop scope iterates and properly drops all values.
+//! No drop-glue is added for types that doesn't need drop. `Scope` allocates enough memory and writes value there, no bookkeeping overhead.
+//!
+//! ## Iterator collecting
+//!
+//! `to_scope_from_iter` method acts as `to_scope` but works on iterators and returns slices.
+//! The limitation is that `to_scope_from_iter` need to allocate memory enough for upper bound of what iterator can yield.
+//! If upper bound is too large or iterator is unbounded it will always fail.
+//! One can use `try_to_scope_from_iter` so fail is `Err` and not panic.
+//! It is safe for iterator to yield more items then upper bound it reports, `to_scope_from_iter` would not iterate past upper bound.
+//! On success it returns mutable reference to slice with items from iterator in order.
+//! All values will be dropped on scope reset or drop, same as with `to_scope`.
+//!
+//! This method is especially useful to deal with API that requires slices (*glares at FFI*), collecting into temporary `Vec` would cost much more.
+//!
 
 #![no_std]
-#![cfg(any(feature = "nightly", feature = "alloc"))]
-#![cfg_attr(feature = "nightly", feature(allocator_api))]
+#![cfg(any(feature = "allocator_api", feature = "alloc"))]
+#![cfg_attr(feature = "allocator_api", feature(allocator_api))]
 
 #[cfg(feature = "alloc")]
 extern crate alloc;
 
+mod allocator_api;
 mod bucket;
 mod drop;
-mod stable_alloc;
 
 use core::{
     alloc::Layout,
@@ -35,149 +146,12 @@ use self::{
     drop::{DropList, WithDrop},
 };
 
-pub use self::stable_alloc::{AllocError, Allocator};
+use self::allocator_api::{AllocError, Allocator};
 
 #[cfg(feature = "alloc")]
-pub use self::stable_alloc::Global;
+use self::allocator_api::Global;
 
-struct OwnedBuckets(Buckets<'static>);
-
-// # Safety
-// OwnedBuckets can be moved to another thread as owned value of `OwnedBuckets` doesn't share references to its interior.
-// OwnedBuckets cannot be `Sync` though.
-unsafe impl Send for OwnedBuckets {}
-
-/// Arena allocator with support for scope based memory reclamation.
-/// This type is a wrapper over underlying allocator
-/// which provides sub-allocations with amortized cost of just few CPU instructions.
-#[cfg(not(feature = "alloc"))]
-pub struct Arena<A: Allocator> {
-    buckets: OwnedBuckets,
-    alloc: A,
-}
-
-/// Arena allocator with support for scope based memory reclamation.
-/// This type is a wrapper over underlying allocator
-/// which provides sub-allocations with amortized cost of just few CPU instructions.
-#[cfg(feature = "alloc")]
-pub struct Arena<A: Allocator = Global> {
-    buckets: OwnedBuckets,
-    alloc: A,
-}
-
-impl<A> Debug for Arena<A>
-where
-    A: Allocator,
-{
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("Arena")
-            .field("buckets", &self.buckets.0)
-            .finish_non_exhaustive()
-    }
-}
-
-impl<A> Drop for Arena<A>
-where
-    A: Allocator,
-{
-    fn drop(&mut self) {
-        unsafe { self.buckets.0.reset(&self.alloc, false) }
-    }
-}
-
-impl<A> Arena<A>
-where
-    A: Allocator,
-{
-    /// Reset pre-allocated memory.
-    /// Keeps only last memory allocated from underlying allocator
-    /// and resets its usage.
-    ///
-    /// This method is safe due to guarantee that all memory allocated
-    /// can't be accessed without some unsafe code breaking safety contract.
-    pub fn reset(&mut self) {
-        unsafe { self.buckets.0.reset(&self.alloc, true) }
-    }
-}
-
-#[cfg(feature = "alloc")]
-impl Arena<Global> {
-    /// Returns new instance of arena allocator based on [`Global`] allocator.
-    pub fn new() -> Self {
-        Arena::new_in(Global)
-    }
-
-    /// Returns new instance of arena allocator based on [`Global`] allocator
-    /// with preallocated capacity in bytes.
-    pub fn with_capacity(capacity: usize) -> Self {
-        Arena::with_capacity_in(capacity, Global)
-    }
-}
-
-impl<A> Arena<A>
-where
-    A: Allocator,
-{
-    /// Returns new instance of arena allocator based on provided allocator.
-    pub fn new_in(alloc: A) -> Self {
-        Arena::with_capacity_in(0, alloc)
-    }
-
-    /// Returns new instance of arena allocator based on provided allocator
-    /// with preallocated capacity in bytes.
-    pub fn with_capacity_in(capacity: usize, alloc: A) -> Self {
-        Arena {
-            buckets: OwnedBuckets(
-                Buckets::new(capacity, &alloc).expect(ALLOCATOR_CAPACITY_OVERFLOW),
-            ),
-            alloc,
-        }
-    }
-
-    /// Allocates a block of memory.
-    /// Returns a [`NonNull<u8>`] meeting the size and alignment guarantees of layout.
-    /// The returned block contents should be considered uninitialized.
-    ///
-    /// Returned block will be deallocated when arena is reset ([`Arena::reset`]) or dropped.
-    #[cfg(all(not(no_global_oom_handling), feature = "alloc"))]
-    pub fn alloc(&self, layout: Layout) -> NonNull<u8> {
-        match self.try_alloc(layout) {
-            Ok(ptr) => ptr,
-            Err(_) => handle_alloc_error(layout),
-        }
-    }
-
-    /// Attempts to allocate a block of memory.
-    /// On success, returns a [`NonNull<u8>`] meeting the size and alignment guarantees of layout.
-    /// The returned block contents should be considered uninitialized.
-    ///
-    /// Returned block will be deallocated when arena is reset ([`Arena::reset`]) or dropped.
-    ///
-    /// # Errors
-    ///
-    /// Returning `Err` indicates that memory is exhausted.
-    pub fn try_alloc(&self, layout: Layout) -> Result<NonNull<u8>, AllocError> {
-        unsafe { self.buckets.0.allocate(layout, &self.alloc) }
-    }
-
-    /// Creates new scope for the arena.
-    /// Returns [`Scope`] instance which can be used to allocate memory and place values into the scope.
-    /// Upon drop, scope will drop all values and free allocated memory.
-    pub fn scope(&mut self) -> Scope<'_, A> {
-        Scope {
-            buckets: self.buckets.0.fork(),
-            alloc: &self.alloc,
-            drop_list: DropList::new(),
-        }
-    }
-
-    /// Reports total memory allocated from underlying allocator by this arena.
-    pub fn total_memory_usage(&self) -> usize {
-        self.buckets.0.total_memory_usage()
-    }
-}
-
-/// Scope associated with `Arena` allocator.
+/// Scope associated with `Scope` allocator.
 /// Allows placing values on the scope returning reference bound to scope borrow.
 /// On drop scope drops all values placed onto it.
 /// On drop scope frees all memory allocated from it.
@@ -188,14 +162,14 @@ pub struct Scope<'arena, A: Allocator> {
     drop_list: DropList<'static>,
 }
 
-/// Scope associated with `Arena` allocator.
+/// Scope associated with `Scope` allocator.
 /// Allows placing values on the scope returning reference bound to scope borrow.
 /// On drop scope drops all values placed onto it.
 /// On drop scope frees all memory allocated from it.
 #[cfg(feature = "alloc")]
 pub struct Scope<'arena, A: Allocator = Global> {
     buckets: Buckets<'arena>,
-    alloc: &'arena A,
+    alloc: A,
     drop_list: DropList<'static>,
 }
 
@@ -217,7 +191,41 @@ where
     fn drop(&mut self) {
         unsafe {
             self.drop_list.reset();
-            self.buckets.reset_fork(self.alloc);
+            self.buckets.reset(&self.alloc, false);
+        }
+    }
+}
+
+#[cfg(feature = "alloc")]
+impl Scope<'_, Global> {
+    /// Returns new instance of arena allocator based on [`Global`] allocator.
+    pub fn new() -> Self {
+        Scope::new_in(Global)
+    }
+
+    /// Returns new instance of arena allocator based on [`Global`] allocator
+    /// with preallocated capacity in bytes.
+    pub fn with_capacity(capacity: usize) -> Self {
+        Scope::with_capacity_in(capacity, Global)
+    }
+}
+
+impl<A> Scope<'_, A>
+where
+    A: Allocator,
+{
+    /// Returns new instance of arena allocator based on provided allocator.
+    pub fn new_in(alloc: A) -> Self {
+        Scope::with_capacity_in(0, alloc)
+    }
+
+    /// Returns new instance of arena allocator based on provided allocator
+    /// with preallocated capacity in bytes.
+    pub fn with_capacity_in(capacity: usize, alloc: A) -> Self {
+        Scope {
+            buckets: Buckets::new(capacity, &alloc).expect(ALLOCATOR_CAPACITY_OVERFLOW),
+            alloc,
+            drop_list: DropList::new(),
         }
     }
 }
@@ -229,7 +237,7 @@ where
     pub fn reset(&mut self) {
         unsafe {
             self.drop_list.reset();
-            self.buckets.reset_fork(self.alloc);
+            self.buckets.reset(&self.alloc, true);
         }
     }
 
@@ -256,7 +264,7 @@ where
     ///
     /// Returning `Err` indicates that memory is exhausted.
     pub fn try_alloc(&self, layout: Layout) -> Result<NonNull<u8>, AllocError> {
-        unsafe { self.buckets.allocate(layout, self.alloc) }
+        unsafe { self.buckets.allocate(layout, &self.alloc) }
     }
 
     /// Move value onto the scope.
@@ -407,21 +415,29 @@ where
         self.buckets.total_memory_usage()
     }
 
-    /// Creates new scope which inherits this scope.
-    /// This scope becomes locked until returned scope is dropped.
-    pub fn scope(&mut self) -> Scope<'_, A> {
-        Scope {
+    /// Creates scope proxy bound to the scope.
+    /// Any objects allocated through proxy will be attached to the scope.
+    /// Returned proxy will use reference to the underlying allocator.
+    pub fn proxy_ref(&mut self) -> ScopeProxy<'_, &'_ A> {
+        ScopeProxy {
             buckets: self.buckets.fork(),
-            alloc: self.alloc,
-            drop_list: DropList::new(),
+            alloc: &self.alloc,
+            drop_list: self.drop_list.fork(),
         }
     }
+}
 
+impl<A> Scope<'_, A>
+where
+    A: Allocator + Clone,
+{
     /// Creates scope proxy bound to the scope.
+    /// Any objects allocated through proxy will be attached to the scope.
+    /// Returned proxy will use clone of the underlying allocator.
     pub fn proxy(&mut self) -> ScopeProxy<'_, A> {
         ScopeProxy {
             buckets: self.buckets.fork(),
-            alloc: self.alloc,
+            alloc: self.alloc.clone(),
             drop_list: self.drop_list.fork(),
         }
     }
@@ -445,7 +461,7 @@ pub struct ScopeProxy<'scope, A: Allocator> {
 #[cfg(feature = "alloc")]
 pub struct ScopeProxy<'scope, A: Allocator = Global> {
     buckets: Buckets<'scope>,
-    alloc: &'scope A,
+    alloc: A,
     drop_list: DropList<'scope>,
 }
 
@@ -499,7 +515,7 @@ where
     ///
     /// Returning `Err` indicates that memory is exhausted.
     pub fn try_alloc(&self, layout: Layout) -> Result<NonNull<u8>, AllocError> {
-        unsafe { self.buckets.allocate(layout, self.alloc) }
+        unsafe { self.buckets.allocate(layout, &self.alloc) }
     }
 
     /// Move value onto the scope.
@@ -650,57 +666,35 @@ where
         self.buckets.total_memory_usage()
     }
 
-    /// Creates new scope which inherits this scope.
+    /// Creates new scope which inherits from the proxy's scope.
     /// This scope becomes locked until returned scope is dropped.
+    /// Returned scope will use reference to the underlying allocator.
+    pub fn scope_ref(&mut self) -> Scope<'_, &'_ A> {
+        Scope {
+            buckets: self.buckets.fork(),
+            alloc: &self.alloc,
+            drop_list: DropList::new(),
+        }
+    }
+}
+
+impl<A> ScopeProxy<'_, A>
+where
+    A: Allocator + Clone,
+{
+    /// Creates new scope which inherits from the proxy's scope.
+    /// This scope becomes locked until returned scope is dropped.
+    /// Returned scope will use clone of the underlying allocator.
     pub fn scope(&mut self) -> Scope<'_, A> {
         Scope {
             buckets: self.buckets.fork(),
-            alloc: self.alloc,
+            alloc: self.alloc.clone(),
             drop_list: DropList::new(),
         }
     }
 }
 
 const ALLOCATOR_CAPACITY_OVERFLOW: &'static str = "Allocator capacity overflow";
-
-unsafe impl<A> Allocator for &'_ Arena<A>
-where
-    A: Allocator,
-{
-    fn allocate(&self, layout: Layout) -> Result<NonNull<[u8]>, AllocError> {
-        let ptr = self.try_alloc(layout)?;
-        let slice = unsafe {
-            // `Arena` always allocates exact amount of bytes.
-            NonNull::new_unchecked(core::slice::from_raw_parts_mut(ptr.as_ptr(), layout.size()))
-        };
-        Ok(slice)
-    }
-
-    unsafe fn deallocate(&self, _ptr: NonNull<u8>, _layout: Layout) {
-        // Will be deallocated on arena reset.
-    }
-
-    #[cfg(feature = "nightly")]
-    unsafe fn shrink(
-        &self,
-        ptr: NonNull<u8>,
-        old_layout: Layout,
-        new_layout: Layout,
-    ) -> Result<NonNull<[u8]>, AllocError> {
-        debug_assert!(
-            new_layout.size() <= old_layout.size(),
-            "`new_layout.size()` must be smaller than or equal to `old_layout.size()`"
-        );
-
-        // Returns same memory unchanged.
-        // This is valid behavior as change in layout won't affect deallocation
-        // and for `grow{_zeroed}` methods new layout with smaller size will only affect numbers of bytes copied.
-        Ok(NonNull::new_unchecked(core::slice::from_raw_parts_mut(
-            ptr.as_ptr(),
-            old_layout.size(),
-        )))
-    }
-}
 
 unsafe impl<A> Allocator for &'_ Scope<'_, A>
 where
@@ -719,7 +713,7 @@ where
         // Will be deallocated on scope drop.
     }
 
-    #[cfg(feature = "nightly")]
+    #[cfg(feature = "allocator_api")]
     unsafe fn shrink(
         &self,
         ptr: NonNull<u8>,
@@ -741,7 +735,7 @@ where
     }
 }
 
-unsafe impl<A> Allocator for &'_ ScopeProxy<'_, A>
+unsafe impl<A> Allocator for ScopeProxy<'_, A>
 where
     A: Allocator,
 {
@@ -758,7 +752,7 @@ where
         // Will be deallocated on scope drop.
     }
 
-    #[cfg(feature = "nightly")]
+    #[cfg(feature = "allocator_api")]
     unsafe fn shrink(
         &self,
         ptr: NonNull<u8>,