+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2021-08-20
+
+### Changed
+- "nightly" feature renamed to "allocator_api" to reflect rust feature it uses.
+- `Arena` type is removed. `Scope` gets methods of the `Arena`.
+
 ## [0.1.1] - 2021-08-19
 
 ### Added